progress made with ORM

+++ src/main/java/com/codecool/pokemoncompanion/repository/AbilityRepository.java
@@ -3,5 +3,5 @@ package com.codecool.pokemoncompanion.repository;
 import com.codecool.pokemoncompanion.model.AbilityWithIdAndPokemonEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AbilityRepository extends JpaRepository<AbilityWithIdAndPokemonEntity, Integer> {
+public interface AbilityRepository extends JpaRepository<AbilityWithIdAndPokemonEntity, Long> {
 }
