+++ src/main/java/com/codecool/pokemoncompanion/repository/TypeRepository.java
@@ -3,5 +3,5 @@ package com.codecool.pokemoncompanion.repository;
 import com.codecool.pokemoncompanion.model.TypeWithIdAndPokemonEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TypeRepository extends JpaRepository<TypeWithIdAndPokemonEntity, Integer> {
+public interface TypeRepository extends JpaRepository<TypeWithIdAndPokemonEntity, Long> {
 }
